@@ -13,51 +13,59 @@ with open(path.join(here, 'README.md'), encoding='utf-8') as f:
 
 setup(
     name='sigmatools',
-    version='0.6',
+    version='0.11',
     description='Tools for the Generic Signature Format for SIEM Systems',
     long_description=long_description,
+    long_description_content_type="text/markdown",
     url='https://github.com/Neo23x0/sigma',
     author='Sigma Project',
     author_email='thomas@patzke.org',
     license='LGPLv3',
     classifiers=[
-        'Development Status :: 3 - Alpha',
+        'Development Status :: 4 - Beta',
         'Intended Audience :: Developers',
         'Intended Audience :: Information Technology',
         'Intended Audience :: System Administrators',
         'Topic :: Security',
         'Topic :: Internet :: Log Analysis',
         'License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
+        'Programming Language :: Python :: 3.7',
         'Environment :: Console',
     ],
     keywords='security monitoring siem logging signatures elasticsearch splunk ids sysmon',
-    packages=['sigma'],
-    python_requires='~=3.5',
-    install_requires=['PyYAML'],
+    packages=['sigma', 'sigma.backends', 'sigma.config', 'sigma.parser'],
+    python_requires='~=3.6',
+    install_requires=['PyYAML', 'pymisp'],
     extras_require={
         'test': ['coverage', 'yamllint'],
     },
     data_files=[
         ('etc/sigma', [
-            'config/arcsight.yml',
-            'config/elk-defaultindex-filebeat.yml',
-            'config/elk-defaultindex-logstash.yml',
+            'config/qualys.yml',
             'config/elk-defaultindex.yml',
-            'config/elk-linux.yml',
+            'config/arcsight.yml',
+            'config/sumologic.yml',
+            'config/netwitness.yml',
             'config/elk-windows.yml',
             'config/helk.yml',
+            'config/elk-defaultindex-logstash.yml',
+            'config/elk-linux.yml',
             'config/logpoint-windows-all.yml',
-            'config/qualys.yml',
-            'config/qradar.yml',
-            'config/spark.yml',
-            'config/netwitness.yml',
+            'config/thor.yml',
+            'config/elk-winlogbeat.yml',
+            'config/elk-defaultindex-filebeat.yml',
             'config/splunk-windows-all.yml',
-            'config/sumologic.yml',
+            'config/qradar.yml',
+            'config/powershell-windows-all.yml',
+            ]),
+        ('etc/sigma/generic', [
+            'config/generic/sysmon.yml',
+            'config/generic/windows-audit.yml',
         ])],
     scripts=[
         'sigmac',
-        'merge_sigma'
+        'merge_sigma',
+        'sigma2misp',
         ]
 )
