@@ -21,8 +21,9 @@ class GraylogQuerystringBackend(SingleTextQueryBackend):
     """Converts Sigma rule into Graylog query string. Only searches, no aggregations."""     
     identifier = "graylog"
     active = True
+    config_required = False
 
-    reEscape = re.compile("([+\\-!(){}\\[\\]^\"~:/]|\\\\(?![*?])|&&|\\|\\|)")
+    reEscape = re.compile("([+\\-!(){}\\[\\]^\"~:/]|(?<!\\\\)\\\\(?![*?\\\\])|&&|\\|\\|)")
     reClear = None
     andToken = " AND "
     orToken = " OR "
