@@ -22,14 +22,16 @@ class GrepBackend(BaseBackend, QuoteCharMixin):
     """Generates Perl compatible regular expressions and puts 'grep -P' around it"""
     identifier = "grep"
     active = True
+    config_required = False
 
-    reEscape = re.compile("([\\|()\[\]{}.^$])")
+    reEscape = re.compile("([\\|()\[\]{}.^$+])")
 
     def generateQuery(self, parsed):
         return "grep -P '^%s'" % self.generateNode(parsed.parsedSearch)
 
     def cleanValue(self, val):
         val = super().cleanValue(val)
+        val = val.replace("'","'\"'\"'")
         return re.sub("\\*", ".*", val)
 
     def generateORNode(self, node):
@@ -51,7 +53,14 @@ class GrepBackend(BaseBackend, QuoteCharMixin):
 
     def generateMapItemNode(self, node):
         key, value = node
-        return self.generateNode(value)
+        if value is None:
+            return self.generateNULLValueNode(node)
+        else:
+            return self.generateNode(value)
 
     def generateValueNode(self, node):
         return self.cleanValue(str(node))
+
+    def generateNULLValueNode(self, node):
+        key, value = node
+        return "(?!%s)" % key
