@@ -14,8 +14,8 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from sigma.parser.condition import ConditionOR
-from .exceptions import SigmaConfigParseError
+from sigma.parser.condition import ConditionOR, NodeSubexpression, ConditionNULLValue
+from .exceptions import SigmaConfigParseError, FieldMappingError
 
 # Field Mapping Definitions
 def FieldMapping(source, target=None):
@@ -47,6 +47,9 @@ class SimpleFieldMapping:
     def resolve_fieldname(self, fieldname):
         return self.target
 
+    def __str__(self):  # pragma: no cover
+        return "SimpleFieldMapping: {} -> {}".format(self.source, self.target)
+
 class MultiFieldMapping(SimpleFieldMapping):
     """1:n field mapping that expands target field names into OR conditions"""
     target_type = list
@@ -56,10 +59,10 @@ class MultiFieldMapping(SimpleFieldMapping):
         cond = ConditionOR()
         for fieldname in self.target:
             cond.add((fieldname, value))
-        return cond
+        return NodeSubexpression(cond)
 
-    def resolve_fieldname(self, fieldname):
-        return self.target
+    def __str__(self):  # pragma: no cover
+        return "MultiFieldMapping: {} -> [{}]".format(self.source, ", ".join(self.target))
 
 class ConditionalFieldMapping(SimpleFieldMapping):
     """
@@ -117,17 +120,110 @@ class ConditionalFieldMapping(SimpleFieldMapping):
                 targets = self.default
 
         if len(targets) == 1:     # result set contains only one target, return mapped item (like SimpleFieldMapping)
-            return (targets.pop(), value)
+            if value is None:
+                return ConditionNULLValue(val=targets.pop())
+            else:
+                return (targets.pop(), value)
         elif len(targets) > 1:        # result set contains multiple targets, return all linked as OR condition (like MultiFieldMapping)
             cond = ConditionOR()
             for target in targets:
-                cond.add((target, value))
-            return cond
+                if value is None:
+                    cond.add(ConditionNULLValue(val=target))
+                else:
+                    cond.add((target, value))
+            return NodeSubexpression(cond)
         else:                       # no mapping found
-            return (key, value)
+            if value is None:
+                return ConditionNULLValue(val=key)
+            else:
+                return (key, value)
 
     def resolve_fieldname(self, fieldname):
         if self.default != None:
             return self.default
         else:
             return fieldname
+
+    def __str__(self):  # pragma: no cover
+        return "ConditionalFieldMapping: {} -> {}".format(self.source, self.target)
+
+# Field mappimg chain
+class FieldMappingChain(object):
+    """
+    Chain of field mappings and fields used for calculation of a field mapping in chained conversion
+    configurations.
+
+    A chain of field mappings may fan out, as one field can map into multiple target fields and these
+    must be propagated further. As the whole chain must be completed at configuration parse time, a
+    restriction applies to conditional field mappings. These are calculated at rule conversion time and
+    therefore it is not possible to decide further mappings after conditionals and these may only appear
+    in the last configuration. This case could be solved by calculation of field mappings at rule conversion
+    time, but it is not considered as important enough to be implemented at this time.
+    """
+    def __init__(self, fieldname):
+        """Initialize field mapping chain with given field name."""
+        self.fieldmappings = set([fieldname])
+
+    def append(self, config):
+        """Propagate current possible field mappings with field mapping from configuration"""
+        if ConditionalFieldMapping in { type(fieldmapping) for fieldmapping in self.fieldmappings }:   # conditional field mapping appeared before, abort.
+            raise FieldMappingError("Conditional field mappings are only allowed in last configuration if configurations are chained.")
+
+        fieldmappings = set()
+        if type(self.fieldmappings) == str:
+            current_fieldmappings = {self.fieldmappings}
+        else:
+            current_fieldmappings = self.fieldmappings
+
+        for fieldname in current_fieldmappings:
+            mapping = config.get_fieldmapping(fieldname)
+            if type(mapping) in (SimpleFieldMapping,  MultiFieldMapping):
+                resolved_mapping = mapping.resolve_fieldname(fieldname)
+                if type(resolved_mapping) is list:
+                    fieldmappings.update(resolved_mapping)
+                else:
+                    fieldmappings.add(resolved_mapping)
+            elif type(mapping) == ConditionalFieldMapping:
+                fieldmappings.add(mapping)
+            else:
+                raise TypeError("Type '{}' is not supported by FieldMappingChain".format(str(type(mapping))))
+
+        if len(fieldmappings) == 1:
+            self.fieldmappings = fieldmappings.pop()
+        else:
+            self.fieldmappings = fieldmappings
+
+    def resolve(self, key, value, sigmaparser):
+        if type(self.fieldmappings) == str:     # one field mapping
+            return (self.fieldmappings, value)
+        elif isinstance(self.fieldmappings, SimpleFieldMapping):
+            return self.fieldmappings.resolve(key, value, sigmaparser)
+        elif type(self.fieldmappings) == set:
+            cond = ConditionOR()
+            for mapping in self.fieldmappings:
+                if type(mapping) == str:
+                    cond.add((mapping, value))
+                elif isinstance(mapping, SimpleFieldMapping):
+                    cond.add(mapping.resolve(key, value, sigmaparser))
+            return NodeSubexpression(cond)
+
+    def resolve_fieldname(self, fieldname):
+        if type(self.fieldmappings) == str:     # one field mapping
+            return self.fieldmappings
+        elif isinstance(self.fieldmappings, SimpleFieldMapping):
+            return self.fieldmappings.resolve_fieldname(fieldname)
+        elif type(self.fieldmappings) == set:
+            mappings = set()
+            for mapping in self.fieldmappings:
+                if type(mapping) == str:
+                    mappings.add(mapping)
+                elif isinstance(mapping, SimpleFieldMapping):
+                    resolved_mapping = mapping.resolve_fieldname(fieldname)
+                    if type(resolved_mapping) is list:
+                        mappings.update(resolved_mapping)
+                    else:
+                        mappings.add(resolved_mapping)
+            return list(mappings)
+
+    def __str__(self):  # pragma: no cover
+        return "FieldMappingChain: {}".format(self.fieldmappings)
